Move most language validation to Python
The error messages produced by the schema validation when using the
<choice><oneOrMore>...</oneOrMore><element>...</element></choice>
syntax are a mess, especially with the optional elements getting in the way.
This commit changes it to only basic validation in the XML schema, and does
the heavy language check lifting in Python.

Requirements on translations are (at least currently) very strict: All
supported languages must be declared, and all translatable text must be
translated into every declared language. The app itself will probably not
be this strict, but it'll probably be a good thing for the repo to enforce.