Wrap list in a top-level object for future-proofing
In case in future we want to add things to this file. If that ever happens I
think it's more likely that we would redesign the whole file and start a v2,
but may as well make it easy to add minor metadata without breaking backwards
compatibility if the need arises.