Split list update to a separate workflow
It was too troublesome to try to get it to commit changes to PRs opened
from forks, and also it's actually not possible.

See "Use in forks from public repositories" in
https://github.com/stefanzweifel/git-auto-commit-action#advanced-uses

So instead, do a dry run on all branches/PRs, but only commit the change
on push to main.